+++ src/gravity/gravity.hpp
@@ -53,7 +53,7 @@ class Gravity {
  private:
   bool haveInitialisedPotential{false};     ///< whether a potential has already been initialised
   bool haveInitialisedBodyForce{false};     ///< whether a body force has already been initialised
-  real centralMass{1.0};                    ///< central mass parameter when central mass potential
+  real centralMass;                    ///< central mass parameter when central mass potential
                                             ///< is enabled
   DataBlock *data;