+++ src/timeIntegrator.hpp
@@ -49,8 +49,8 @@ class TimeIntegrator {
   int64_t ncycles;        // # of cycles
   double lastLog;         // time for the last log (s)
   double lastMpiLog;      // time for the last MPI log (s)
-  double maxRuntime{-1.0};      // Maximum runtime requested (disabled when negative)
-  int64_t cyclePeriod = 100;    // # of cycles between two logs
+  double maxRuntime;      // Maximum runtime requested (disabled when negative)
+  int64_t cyclePeriod;    // # of cycles between two logs
   Kokkos::Timer timer;    // Internal timer of the integrator
 };