+++ test/HD/FargoPlanet/setup.cpp
@@ -167,13 +167,13 @@ Setup::Setup(Input &input, Grid &grid, DataBlock &data, Output &output)// : m_pl
   data.hydro.viscosity.EnrollViscousDiffusivity(&MyViscosity);
   if(data.haveFargo)
     data.fargo.EnrollVelocity(&FargoVelocity);
-  sigma0Glob = input.GetReal("Setup","sigma0",0);
-  sigmaSlopeGlob = input.GetReal("Setup","sigmaSlope",0);
-  h0Glob = input.GetReal("Setup","h0",0);
-  alphaGlob = input.GetReal("Setup","alpha",0);
-  qpGlob = input.GetReal("Planet","qpl",0);
-  RpGlob = input.GetReal("Planet","Rpl",0);
-  thicknessSmoothingGlob = input.GetReal("Planet","thicknessSmoothing",0);
+  sigma0Glob = input.Get<real>("Setup","sigma0",0);
+  sigmaSlopeGlob = input.Get<real>("Setup","sigmaSlope",0);
+  h0Glob = input.Get<real>("Setup","h0",0);
+  alphaGlob = input.Get<real>("Setup","alpha",0);
+  qpGlob = input.Get<real>("Planet","qpl",0);
+  RpGlob = input.Get<real>("Planet","Rpl",0);
+  thicknessSmoothingGlob = input.Get<real>("Planet","thicknessSmoothing",0);
   idfx::cout << "alpha= " << alphaGlob << std::endl;
 
 }