+++ test/MHD/ShearingBox/setup.cpp
@@ -71,10 +71,10 @@ Setup::Setup(Input &input, Grid &grid, DataBlock &data, Output &output) {
   gammaIdeal=data.hydro.GetGamma();
 
   // Get rotation rate along vertical axis
-  omega=input.GetReal("Hydro","rotation",0);
-  shear=input.GetReal("Hydro","shearingBox",0);
-  B0y = input.GetReal("Setup","B0y",0);
-  B0z = input.GetReal("Setup","B0z",0);
+  omega=input.Get<real>("Hydro","rotation",0);
+  shear=input.Get<real>("Hydro","shearingBox",0);
+  B0y = input.Get<real>("Setup","B0y",0);
+  B0z = input.Get<real>("Setup","B0z",0);
 
   // Add our userstep to the timeintegrator
   data.gravity.EnrollBodyForce(BodyForce);